+++ src/services/ClusterService.ts
@@ -1,3 +1,4 @@
+import type { ICloudProvider } from "@/entities";
 import type { ICluster } from "@/entities/Cluster";
 import { clusterSchema } from "@/entities/Cluster";
 import type { IQueryFilter, IQueryOptions } from "@/interfaces";
@@ -12,6 +13,38 @@ import BaseService from "./BaseService";
 export class ClusterService extends BaseService<ICluster> {
 	constructor(ownership?: Ownership) {
 		super(clusterSchema, ownership);
+	}
+
+	async updateOne(filter: IQueryFilter<ICluster>, data: any, options?: IQueryOptions): Promise<ICluster> {
+		let cluster = await this.findOne(filter, { ...options, populate: ["provider"] });
+		if (!cluster) {
+			if (filter.owner) {
+				throw new Error(`Unauthorized.`);
+			} else {
+				throw new Error(`Cluster not found.`);
+			}
+		}
+
+		// get cloud provider of this cluster
+		const cloudProvider = cluster.provider as ICloudProvider;
+		if (!cloudProvider) throw new Error(`Cloud Provider not found.`);
+
+		// validation - check cluster accessibility
+		if (cloudProvider.shortName === "gcloud") {
+			if (!cluster.serviceAccount && !data.serviceAccount) throw new Error(`Google Service Account (JSON) is required.`);
+			if (!cluster.zone && !data.zone) throw new Error(`Google cluster zone is required.`);
+		}
+		if (cloudProvider.shortName === "digitalocean") {
+			if (!cluster.apiAccessToken) throw new Error(`Digital Ocean API Access Token is required.`);
+		}
+		if (cloudProvider.shortName === "custom") {
+			if (!cluster.kubeConfig && !data.kubeConfig) throw new Error(`Kube config data (YAML) is required.`);
+		}
+
+		// update to database
+		cluster = await super.updateOne({ _id: cluster._id }, data, options);
+
+		return cluster;
 	}
 
 	async delete(filter?: IQueryFilter<ICluster>, options?: IQueryOptions): Promise<{ ok: boolean; affected: number }> {